requirements.txt
certifi==2020.12.5
click==7.1.2
cloudpickle==1.6.0
joblib==1.0.0
llvmlite==0.35.0
numba==0.52.0
numpy==1.19.4
pandas==1.2.0
pyarrow==2.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
scikit-learn==0.24.0
scipy==1.6.0
shap @ git+https://github.com/slundberg/shap@6e04be3f4a9c7d61ada301366e4273a096e5c1f5
six==1.15.0
slicer==0.0.7
threadpoolctl==2.1.0
tqdm==4.55.0
